arth-executive/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
pydantic-settings>=2.2.0
httpx>=0.27.0
